Move routes constants to navigation lib
