+++ client/src/routes/Sessions/components/SessionModal.tsx
@@ -19,7 +19,7 @@ import HalfModal from '../../../common/components/Modals/HalfModal';
 import {Spacer16, Spacer8} from '../../../common/components/Spacers/Spacer';
 import {Display24} from '../../../common/components/Typography/Display/Display';
 import {COLORS} from '../../../../../shared/src/constants/colors';
-import {RootStackProps} from '../../../common/constants/routes';
+import {RootStackProps} from '../../../lib/navigation/constants/routes';
 import useExerciseById from '../../../lib/content/hooks/useExerciseById';
 import NS from '../../../lib/i18n/constants/namespaces';
 import {userAtom} from '../../../lib/user/state/state';
